remove fast_log.h
